% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_lambda.R
\name{fixed_lambda}
\alias{fixed_lambda}
\title{Fits a HIP model given lambda values}
\usage{
fixed_lambda(X, Y, gamma, family, topn, lambda_xi=1, lambda_g=1, K=NULL,
k_thresh=0.2, update_thresh=10^-5, epsilon=10^-4, max_iter=50,
standardize='subgroup', std_type='scale_center', std_x=TRUE,
std_y=TRUE, verbose=FALSE)
}
\arguments{
\item{X}{matrix list - list of X^d,s matrices containing covariates}

\item{Y}{matrix list - list of Y^s matrices containing outcomes}

\item{family}{string - family of outcome; options are 'gaussian', 'multiclass', 'poisson' or 'zip'}

\item{topn}{int or list\if{html}{\out{<int>}} - number of variables to retain; different values may be specified for each view using a list of length D}

\item{lambda_xi}{double - value of lambda_xi in penalty term}

\item{lambda_g}{double - value of lambda_g in penalty term}

\item{K}{int - number of latent components; will use 'select_K_simple' to choose if not provided}

\item{k_thresh}{double - threshold to use for 'select_K_simple' if K not provided; default = 0.2}

\item{update_thresh}{double - threshold to use for suboptimization convergence criterion; default = 10**-5}

\item{epsilon}{double - threshold to use for overall convergence criterion; default = 10**-4}

\item{max_iter}{int - maximum number of outer loop iterations; default = 50}

\item{standardize}{string - One of "all", "subgroup", or "none"; default = "subgroup"}

\item{std_type}{string - One of "scale_center", "center", or "norm"; default = "scale_center"}

\item{std_x}{boolean - indicates whether to standardize X; default = TRUE}

\item{std_y}{boolean - indicates whether to standardize Y; default = TRUE}

\item{verbose}{boolean - whether to print additional information during optimization; default = FALSE}
}
\value{
\item{\code{full}}{list - results returned from optimize_torch on full data }
\item{\code{include}}{list\if{html}{\out{<tensor>}} - list of length D with a 1 indicating the variable
was included in subset fit and 0 indicating not included in subset fit }
\item{\code{subset} }{list - results returned from optimize_torch on subset of variables}

In addition, \code{fixed_lambda} returns -99 for \code{best_index} and NA for \code{criterion}
to indicate they are not applicable to this case. topn is also returned.
}
\description{
\code{fixed_lambda} fits a model from X, Y data and fixed \eqn{\lambda_\xi}, \eqn{\lambda_G}
supplied by the user.
}
\examples{
# Generating data
dat_gen <- generate_data()

# Getting model from fixed_lambda
res <- fixed_lambda(dat_gen$X, dat_gen$Y, c(1,1), 'gaussian', 50)

}
