% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_R.R
\name{get_best_r}
\alias{get_best_r}
\title{Helper function to get best result from select_lambda}
\usage{
get_best_r(res_list, criterion="", family)
}
\arguments{
\item{res_list}{list - search_results list from the select_lambda output}

\item{criterion}{string - criterion for model selection: one of 'CV', 'BIC', 'AIC', 'eBIC_0', 'eBIC_5', or 'eBIC_1'}

\item{family}{string - family of outcome data ('gaussian', 'multiclass', etc.)}
}
\value{
\code{get_best_R} returns the index of the best results from the select_lambda output
}
\description{
'get_best_R' is used within the select_lambda function to get the best result based on
model performance. Not generally called by the user.
}
\details{
get_best_R takes in the list of results from the select_lambda tuning function, and
returns the index of the best result. It also takes in the criterion used in training
such as AIC or BIC, in addition to the family the outcome data belongs to.
}
\examples{
# Generate data
dat_gen <- generate_data()

# Get results from select_lambda
res <- select_lambda(dat_gen$X, dat_gen$Y, c(1,1), "gaussian", 50, K=2, num_steps=c(4,4))

best_index <- get_best_r(res$out$search_results, "eBIC_0", "gaussian")

}
