% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_plots.R
\name{variable_plots}
\alias{variable_plots}
\title{Variable importance plots}
\usage{
variable_plots(data, res, fix_or_search, top_plotted=15, compact=TRUE,
output_table=FALSE, file_path=NA)
}
\arguments{
\item{data}{data object from \code{generate_data} or \code{format_data}}

\item{res}{output from \code{select_lambda} or \code{fixed_lambda}}

\item{fix_or_search}{string - 'fixed' if results obtained from \code{fixed_lambda}, 'search' otherwise}

\item{top_plotted}{int - number of variables to include in each variable importance plot; 15 by default}

\item{compact}{boolean - if TRUE (default), only plots top_plotted variables; if set to FALSE, plots as many as possible}

\item{output_table}{boolean - set to TRUE to write variable importance results to an Excel file; FALSE by default}

\item{file_path}{string - if output_table is TRUE, you may also specify a file path and file name; make sure to add the .xlsx extension to your file name.
By default, it will be written to your current working directory with the name "variable_importance_table.xlsx"}
}
\value{
Returns the variable importance plot as a ggplot object
}
\description{
Outputs plots showing the highest ranking variables contributing to model performance.
Plots are shown for each data view and subgroup.
}
\examples{

# Generate data
dat_gen <- generate_data()

# Get results from select_lambda
res <- select_lambda(dat_gen$X, dat_gen$Y, c(1,1), 'gaussian', 50)

# Variable importance plots
variable_plots(dat_gen, res, 'search', output_table=TRUE)

}
